requirements.txt
openai
transformers
accelerate
gradio
python-dotenv